Some extra checks

+++ server/routes/user.js
@@ -5,23 +5,38 @@ const bcrypt = require("bcryptjs")
 
 // GET A USER BY ID
 router.get("/:id", (req, res) => {
+    // the request header has the token then we can verify it
+    if (!req.headers.authorization) {
+        return res.status(401).json({
+            message: "Unauthorized"
+        })
+    }
     // get the user if the user has valid token 
     const token = req.headers.authorization.split(" ")[1]
-    jwt.verify(token, process.env.SECRET_KEY, (err, decoded) => { 
-        if(err) {
+
+    jwt.verify(token, process.env.SECRET_KEY, (err, decoded) => {
+        if (err) {
             return res.status(401).json({
                 message: "Unauthorized"
             })
         }
         user.findById(req.params.id).then(user => {
-            if(!user) {
+            if (!user) {
                 return res.status(404).json({
                     message: "User not found"
                 })
             }
+            // if the user is found, do not send the password and timestamp
+            const {
+                password,
+                updsatedAt,
+                ...other
+            } = user._doc
             res.status(200).json({
-                user
+                message: "User found",
+                user: other
             })
+
         }).catch(err => {
             res.status(500).json({
                 message: "Error getting user"
@@ -33,26 +48,41 @@ router.get("/:id", (req, res) => {
 
 // UPDATE A USER BY ID
 router.put("/:id", (req, res) => {
+    // the request header has the token then we can verify it
+    if (!req.headers.authorization) {
+        return res.status(401).json({
+            message: "Unauthorized"
+        })
+    }
     // get the user if the user has valid token 
     const token = req.headers.authorization.split(" ")[1]
-    jwt.verify(token, process.env.SECRET_KEY, (err, decoded) => { 
-        if(err) {
+    jwt.verify(token, process.env.SECRET_KEY, (err, decoded) => {
+        if (err) {
             return res.status(401).json({
                 message: "Unauthorized"
             })
         }
         // if the user is updating their password then hash it 
-        if (req.body.password) { 
+        if (req.body.password) {
             req.body.password = bcrypt.hashSync(req.body.password, 16)
         }
-        user.findByIdAndUpdate(req.params.id, req.body, { new: true }).then(user => { 
-            if(!user) {
+        user.findByIdAndUpdate(req.params.id, req.body, {
+            new: true
+        }).then(user => {
+            if (!user) {
                 return res.status(404).json({
                     message: "User not found"
                 })
             }
+            // if the user is found, do not send the password and timestamp
+            const {
+                password,
+                updatedAt,
+                ...userData
+            } = user._doc
             res.status(200).json({
-                user
+                message: "User updated",
+                user: userData
             })
         }).catch(err => {
             res.status(500).json({
@@ -65,16 +95,22 @@ router.put("/:id", (req, res) => {
 
 // DELETE A USER BY ID
 router.delete("/:id", (req, res) => {
+    // the request header has the token then we can verify it
+    if (!req.headers.authorization) {
+        return res.status(401).json({
+            message: "Unauthorized"
+        })
+    }
     // get the user if the user has valid token 
     const token = req.headers.authorization.split(" ")[1]
-    jwt.verify(token, process.env.SECRET_KEY, (err, decoded) => { 
-        if(err) {
+    jwt.verify(token, process.env.SECRET_KEY, (err, decoded) => {
+        if (err) {
             return res.status(401).json({
                 message: "Unauthorized"
             })
         }
         user.findByIdAndDelete(req.params.id).then(user => {
-            if(!user) {
+            if (!user) {
                 return res.status(404).json({
                     message: "User not found"
                 })
